Code seperation and new command
I removed the code from the main class and put it in their respective folders. Added a new command to get all user(crashes and does not work yet). Added an ID's class so I can reference that to get a certain variable rather than repeatedly remaking that one variable.